pyproject.toml
[project]
name = "vision-forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "python-multipart>=0.0.27",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.46.0",
]
